fastapi/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "ollama",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langchain-ollama",
    "langgraph",
    "llama-index",
    "torch>=2.1.0",
    "transformers==4.57.3",
    "sentence-transformers==3.0.1",
    "opik",
    "loguru",
    'twikit',
    'linkedin-api',
    "pydantic",
    "crawl4ai",
    "python-dotenv",
    "qdrant-client",
    "numpy>=1.24.4,<2.0",
    "pymongo",
    "selenium",
    "chromedriver-autoinstaller",
    "brotli",
    "html2text",
    "markdownify",
    "beautifulsoup4>=4.14.3",
    "tavily-python",
    "ddgs",
    "wikipedia>=1.4.0",
    "sqlmodel>=0.0.32",
    "duckduckgo-search>=8.1.1",
    "datasets>=4.5.0",
    "instructor",
    "soundfile",
    "grandalf",
    "chromadb>=1.4.1",
    "langsmith",
    "TTS>=0.22.0",
    "torchaudio>=2.1.0",
    "melotts @ git+https://github.com/myshell-ai/MeloTTS.git",
    "librosa==0.9.1",
    "eng_to_ipa",
    "inflect",
    "unidecode",
    "pypinyin",
    "cn2an",
    "jieba",
    "langid",
    "faster-whisper",
    "whisper-timestamped",
    "pydub",
    "wavmark",
    "qwen-tts>=0.1.1",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.12",
]

[tool.uv]
override-dependencies = [
    "langsmith>=0.1.0, <0.4.0",
    "transformers==4.57.3",
    "numpy>=1.24.4,<2.0",
    "librosa>=0.9.1"
]
